+++ pkg/refresh/refresh.go
@@ -95,6 +95,11 @@ func (s *InstanceRefreshService) Refresh(ctx context.Context, minHealthyPercenta
 			}
 		}
 
+		if len(asg.Instances) == 0 {
+			s.Scope.Info(fmt.Sprintf("ASG %s has no instances, skipping...", *asg.AutoScalingGroupName))
+			continue
+		}
+
 		refreshInput := &autoscaling.StartInstanceRefreshInput{
 			AutoScalingGroupName: asg.AutoScalingGroupName,
 			DesiredConfiguration: &autoscaling.DesiredConfiguration{